Move EventSelector to own file
in same directory
